# x-pack/filebeat/input/httpjson: remove concurrency from response processing