<!-- Type of change
Please label this PR with one of the following labels, depending on the scope of your change:
- Bug
- Enhancement
- Breaking change
- Deprecation
- Cleanup
- Docs
-->

## Proposed commit message

<!-- Mandatory
Explain here the changes you made on the PR.

Please explain:

- WHAT: patterns used, algorithms implemented, design architecture, message processing, etc.
- WHY:  the rationale/motivation for the changes

This text will be pasted into the squash dialog when the change is committed and will be
a long term historical record of the change to help future contributors understand the
change, please help them by making it clear and comprehensive, they may be you.

If the commit title is adequate to describe both of these things, The text here may be omitted
or replaced with "See title". The title of the PR will be used as the commit message title when
the merge is made and the "See title" marker will be removed if present.

The text here and the PR title will be subject to the PR review process.
-->

Replace concurrent code and locks with sequential processing using handler interface types to simplify data flow.

Post refactor clean up will happen in a follow-up PR.

## Checklist

<!-- Mandatory
Add a checklist of things that are required to be reviewed in order to have the PR approved

List here all the items you have verified BEFORE sending this PR. Please DO NOT remove any item, striking through those that do not apply. (Just in case, strikethrough uses two tildes. ~~Scratch this.~~)
-->

- [x] My code follows the style guidelines of this project
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [ ] I have made corresponding change to the default configuration files
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] I have added an entry in `CHANGELOG.next.asciidoc` or `CHANGELOG-developer.next.asciidoc`.

## Author's Checklist

<!-- Recommended
Add a checklist of things that are required to be reviewed in order to have the PR approved
-->
- [ ] This PR is subtle. I recommend review by commit. It is a little wandery (I noticed some things a commit after I should have), but the logic of the progression is fairly straightforward if the changes are stepped through rather than swallowed whole.
- [ ] Ignore linter complaints; they will be repaired in the follow-up.

## How to test this PR locally

<!-- Recommended
Explain here how this PR will be tested by the reviewer: commands, dependencies, steps, etc.
-->

## Related issues

<!-- Recommended
Link related issues below. Insert the issue link or reference after the word "Closes" if merging this should automatically close it.

- Closes #123
- Relates #123
- Requires #123
- Superseds #123
-->
- For #35432

## Use cases

<!-- Recommended
Explain here the different behaviors that this PR introduces or modifies in this project, user roles, environment configuration, etc.

If you are familiar with Gherkin test scenarios, we recommend its usage: https://cucumber.io/docs/gherkin/reference/
-->

## Screenshots

<!-- Optional
Add here screenshots about how the project will be changed after the PR is applied. They could be related to web pages, terminal, etc, or any other image you consider important to be shared with the team.
-->

## Logs

<!-- Recommended
Paste here output logs discovered while creating this PR, such as stack traces or integration logs, or any other output you consider important to be shared with the team.
-->

## Performance impact

The change here is not intended to be a performance improvement, but rather to improve the maintainability and correctness of the code. However, it's worth checking that it does not do too much harm in that regard. The following is a comparison of the parent of this set of changes and the change here.

On mac there is little to no change in the cpu bench performance of the package after the change and a small but consistent improvement in the memory bench performance. A single cpu bench does show a significant improvement, and this is consistent over runs, but is not an important bench and is probably just a weirdness.

```
goos: darwin
goarch: amd64
pkg: github.com/elastic/beats/v7/x-pack/filebeat/input/httpjson
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
                                                                                                        │   old.bench    │                new.bench                 │
                                                                                                        │     sec/op     │     sec/op      vs base                  │
Input/simple_GET_request-16                                                                                547.5µ ±  19%    523.0µ ±  27%        ~ (p=0.739 n=10)
Input/simple_HTTPS_GET_request-16                                                                          3.193m ±  15%    3.283m ±  11%        ~ (p=0.165 n=10)
Input/request_honors_rate_limit-16                                                                         628.2µ ±  55%    635.2µ ±  17%        ~ (p=0.447 n=9+10)
Input/request_retries_when_failed-16                                                                       872.9µ ± 213%    824.3µ ±  31%        ~ (p=0.604 n=10+9)
Input/POST_request_with_body-16                                                                           1287.3µ ± 117%    881.5µ ±  90%        ~ (p=0.143 n=10)
Input/repeated_POST_requests-16                                                                            103.2m ±  43%    103.4m ±  74%        ~ (p=0.393 n=10)
Input/split_by_json_objects_array-16                                                                       882.4µ ±  67%    932.9µ ±  37%        ~ (p=0.912 n=10)
Input/split_by_json_objects_array_with_keep_parent-16                                                      848.4µ ±  11%    925.5µ ± 172%        ~ (p=0.315 n=9+10)
Input/split_on_empty_array_without_ignore_empty_value-16                                                   1.129m ±  71%    1.358m ± 159%        ~ (p=0.631 n=10)
Input/split_on_empty_array_with_ignore_empty_value-16                                                       5.001 ±   0%     5.001 ±   0%        ~ (p=0.739 n=10)
Input/split_on_null_field_with_ignore_empty_value_keeping_parent-16                                        555.4µ ±  23%    519.0µ ±  31%        ~ (p=0.579 n=10)
Input/split_on_empty_array_with_ignore_empty_value_keeping_parent-16                                       918.9µ ±  12%    894.8µ ±  11%        ~ (p=0.968 n=9+10)
Input/split_on_null_field_at_root_with_ignore_empty_value_keeping_parent-16                                805.2µ ±  18%    798.5µ ±  29%        ~ (p=0.968 n=10+9)
Input/split_on_empty_array_at_root_with_ignore_empty_value_keeping_parent-16                               922.8µ ±  25%    715.8µ ±  19%  -22.43% (p=0.001 n=9+10)
Input/nested_split-16                                                                                      1.243m ±  59%    1.020m ±  37%        ~ (p=0.762 n=10+8)
Input/split_events_by_not_found-16                                                                          5.000 ±   0%     5.000 ±   0%        ~ (p=0.631 n=10)
Input/pagination-16                                                                                        6.838m ±  14%    6.763m ±   5%        ~ (p=0.912 n=10)
Input/first_event-16                                                                                        2.004 ±   0%     2.005 ±   5%        ~ (p=0.315 n=10)
Input/oauth2-16                                                                                            1.089m ±  33%    1.032m ±  29%        ~ (p=0.912 n=10)
Input/request_transforms_can_access_state_from_previous_transforms-16                                      1.177m ±   7%    1.119m ±  12%        ~ (p=0.247 n=10)
Input/response_transforms_can't_access_request_state_from_previous_transforms-16                           882.8µ ±  13%    943.2µ ±   9%        ~ (p=0.113 n=9)
Input/simple_Chain_GET_request-16                                                                          1.603m ±  24%    1.449m ±  25%        ~ (p=0.739 n=10)
Input/multiple_Chain_GET_request-16                                                                        6.836m ± 249%    4.063m ± 120%        ~ (p=0.243 n=10+9)
Input/date_cursor_while_using_chain-16                                                                     2.050m ±  37%    1.665m ± 107%        ~ (p=0.143 n=10)
Input/split_by_json_objects_array_in_chain-16                                                              2.991m ±  64%    2.718m ±  47%        ~ (p=0.631 n=10)
Input/split_by_json_objects_array_with_keep_parent_in_chain-16                                             2.439m ±  44%    2.219m ±  50%        ~ (p=0.796 n=10)
Input/pagination_when_used_with_chaining-16                                                                5.331m ±  26%    4.548m ±  80%        ~ (p=0.684 n=10)
Input/replace_with_clause_and_first_response_object-16                                                     4.942m ± 207%    7.792m ±  61%        ~ (p=1.000 n=10)
Input/replace_with_clause_with_hardcoded_value_1-16                                                        3.493m ±  57%    4.563m ±  50%        ~ (p=0.436 n=10)
Input/replace_with_clause_with_hardcoded_value_(no_dot_prefix)-16                                          4.746m ±  56%    4.400m ± 145%        ~ (p=0.739 n=10)
Input/replace_with_clause_with_hardcoded_value_(more_than_one_dot_prefix)-16                               3.479m ±  98%    4.548m ± 197%        ~ (p=0.436 n=10)
Input/replace_with_clause_with_hardcoded_value_containing_'.'_(dots)-16                                    4.249m ±  77%    4.641m ± 106%        ~ (p=0.971 n=10)
Input/global_transform_context_separation_with_parent_last_response_object-16                              8.644m ±  57%    7.546m ± 126%        ~ (p=0.631 n=10)
Input/cursor_value_is_updated_for_root_response_with_chaining_&_pagination-16                              10.65m ±  32%    13.54m ±  48%        ~ (p=0.247 n=10)
Input/cursor_value_is_updated_for_root_response_with_chaining_&_pagination_along_with_split_operator-16    9.896m ±  58%   10.520m ±  52%        ~ (p=1.000 n=10)
Input/Test_simple_XML_decode-16                                                                            14.48m ±  69%    14.91m ±  33%        ~ (p=0.796 n=10)
geomean                                                                                                    4.478m           4.392m          -1.92%

                                                                                                        │  old.bench   │               new.bench                │
                                                                                                        │     B/op     │     B/op       vs base                 │
Input/simple_GET_request-16                                                                               99.97Ki ± 0%    99.34Ki ± 0%  -0.63% (p=0.001 n=10)
Input/simple_HTTPS_GET_request-16                                                                         182.6Ki ± 0%    181.8Ki ± 0%  -0.46% (p=0.001 n=10)
Input/request_honors_rate_limit-16                                                                        98.91Ki ± 0%    97.84Ki ± 0%  -1.08% (p=0.000 n=9+10)
Input/request_retries_when_failed-16                                                                      99.70Ki ± 0%    98.89Ki ± 1%  -0.81% (p=0.013 n=10+9)
Input/POST_request_with_body-16                                                                           101.1Ki ± 0%    100.4Ki ± 0%  -0.69% (p=0.001 n=10)
Input/repeated_POST_requests-16                                                                           136.9Ki ± 2%    134.7Ki ± 1%       ~ (p=0.063 n=10)
Input/split_by_json_objects_array-16                                                                      103.3Ki ± 0%    102.2Ki ± 1%  -1.08% (p=0.000 n=10)
Input/split_by_json_objects_array_with_keep_parent-16                                                     103.9Ki ± 0%    103.0Ki ± 0%  -0.94% (p=0.000 n=9+10)
Input/split_on_empty_array_without_ignore_empty_value-16                                                  99.79Ki ± 0%   100.04Ki ± 0%  +0.25% (p=0.002 n=10)
Input/split_on_empty_array_with_ignore_empty_value-16                                                     235.8Ki ± 4%    224.9Ki ± 5%  -4.62% (p=0.007 n=10)
Input/split_on_null_field_with_ignore_empty_value_keeping_parent-16                                       99.46Ki ± 0%    98.87Ki ± 0%  -0.59% (p=0.001 n=10)
Input/split_on_empty_array_with_ignore_empty_value_keeping_parent-16                                      99.69Ki ± 0%    99.30Ki ± 0%  -0.39% (p=0.001 n=9+10)
Input/split_on_null_field_at_root_with_ignore_empty_value_keeping_parent-16                               99.09Ki ± 0%    98.76Ki ± 0%  -0.34% (p=0.002 n=10+9)
Input/split_on_empty_array_at_root_with_ignore_empty_value_keeping_parent-16                              99.24Ki ± 0%    98.55Ki ± 0%  -0.70% (p=0.001 n=9+10)
Input/nested_split-16                                                                                     104.0Ki ± 0%    103.1Ki ± 0%  -0.87% (p=0.004 n=10+8)
Input/split_events_by_not_found-16                                                                        235.8Ki ± 5%    232.1Ki ± 6%       ~ (p=0.631 n=10)
Input/pagination-16                                                                                       420.1Ki ± 0%    417.6Ki ± 0%  -0.60% (p=0.001 n=10)
Input/first_event-16                                                                                      320.9Ki ± 4%    315.8Ki ± 6%       ~ (p=0.631 n=10)
Input/oauth2-16                                                                                           124.9Ki ± 0%    124.3Ki ± 0%  -0.53% (p=0.001 n=10)
Input/request_transforms_can_access_state_from_previous_transforms-16                                     161.5Ki ± 0%    160.8Ki ± 0%  -0.44% (p=0.002 n=10)
Input/response_transforms_can't_access_request_state_from_previous_transforms-16                          142.3Ki ± 0%    141.3Ki ± 0%  -0.69% (p=0.000 n=9)
Input/simple_Chain_GET_request-16                                                                         160.1Ki ± 0%    159.0Ki ± 0%  -0.73% (p=0.001 n=10)
Input/multiple_Chain_GET_request-16                                                                       216.8Ki ± 1%    215.2Ki ± 1%  -0.75% (p=0.002 n=10+9)
Input/date_cursor_while_using_chain-16                                                                    210.8Ki ± 0%    209.8Ki ± 0%  -0.45% (p=0.001 n=10)
Input/split_by_json_objects_array_in_chain-16                                                             163.7Ki ± 0%    162.1Ki ± 0%  -1.00% (p=0.000 n=10)
Input/split_by_json_objects_array_with_keep_parent_in_chain-16                                            164.2Ki ± 0%    162.6Ki ± 0%  -0.99% (p=0.001 n=10)
Input/pagination_when_used_with_chaining-16                                                               251.5Ki ± 0%    247.5Ki ± 0%  -1.58% (p=0.000 n=10)
Input/replace_with_clause_and_first_response_object-16                                                    245.8Ki ± 1%    244.1Ki ± 0%  -0.69% (p=0.001 n=10)
Input/replace_with_clause_with_hardcoded_value_1-16                                                       187.3Ki ± 1%    185.8Ki ± 0%  -0.80% (p=0.001 n=10)
Input/replace_with_clause_with_hardcoded_value_(no_dot_prefix)-16                                         190.4Ki ± 0%    188.6Ki ± 1%  -0.95% (p=0.000 n=10)
Input/replace_with_clause_with_hardcoded_value_(more_than_one_dot_prefix)-16                              191.6Ki ± 0%    190.0Ki ± 1%  -0.84% (p=0.000 n=10)
Input/replace_with_clause_with_hardcoded_value_containing_'.'_(dots)-16                                   187.8Ki ± 1%    186.1Ki ± 0%  -0.86% (p=0.001 n=10)
Input/global_transform_context_separation_with_parent_last_response_object-16                             363.9Ki ± 1%    359.2Ki ± 0%  -1.29% (p=0.000 n=10)
Input/cursor_value_is_updated_for_root_response_with_chaining_&_pagination-16                             447.2Ki ± 0%    447.7Ki ± 1%       ~ (p=0.684 n=10)
Input/cursor_value_is_updated_for_root_response_with_chaining_&_pagination_along_with_split_operator-16   459.7Ki ± 0%    456.9Ki ± 1%  -0.60% (p=0.017 n=10)
Input/Test_simple_XML_decode-16                                                                           2.644Mi ± 0%    2.644Mi ± 0%       ~ (p=0.739 n=10)
geomean                                                                                                   180.9Ki         179.3Ki       -0.88%

                                                                                                        │  old.bench  │              new.bench               │
                                                                                                        │  allocs/op  │  allocs/op   vs base                 │
Input/simple_GET_request-16                                                                                634.0 ± 0%    627.0 ± 0%  -1.10% (p=0.000 n=10)
Input/simple_HTTPS_GET_request-16                                                                         1.697k ± 0%   1.690k ± 0%  -0.41% (p=0.000 n=10)
Input/request_honors_rate_limit-16                                                                         605.0 ± 0%    597.0 ± 0%  -1.32% (p=0.000 n=9+10)
Input/request_retries_when_failed-16                                                                       605.0 ± 0%    598.0 ± 0%  -1.16% (p=0.001 n=10+9)
Input/POST_request_with_body-16                                                                            643.0 ± 0%    636.0 ± 0%  -1.09% (p=0.000 n=10)
Input/repeated_POST_requests-16                                                                            924.0 ± 1%    908.0 ± 1%  -1.73% (p=0.000 n=10)
Input/split_by_json_objects_array-16                                                                       658.0 ± 0%    646.0 ± 0%  -1.82% (p=0.000 n=10)
Input/split_by_json_objects_array_with_keep_parent-16                                                      668.0 ± 0%    656.0 ± 0%  -1.80% (p=0.000 n=9+10)
Input/split_on_empty_array_without_ignore_empty_value-16                                                   615.0 ± 0%    610.0 ± 0%  -0.81% (p=0.000 n=10)
Input/split_on_empty_array_with_ignore_empty_value-16                                                     1.526k ± 6%   1.521k ± 5%       ~ (p=0.288 n=10)
Input/split_on_null_field_with_ignore_empty_value_keeping_parent-16                                        613.0 ± 0%    606.0 ± 0%  -1.14% (p=0.000 n=10)
Input/split_on_empty_array_with_ignore_empty_value_keeping_parent-16                                       615.0 ± 0%    608.0 ± 0%  -1.14% (p=0.000 n=9+10)
Input/split_on_null_field_at_root_with_ignore_empty_value_keeping_parent-16                                610.0 ± 0%    603.0 ± 0%  -1.15% (n=10+9)
Input/split_on_empty_array_at_root_with_ignore_empty_value_keeping_parent-16                               612.0 ± 0%    605.0 ± 0%  -1.14% (p=0.000 n=9+10)
Input/nested_split-16                                                                                      665.0 ± 0%    652.0 ± 0%  -1.95% (n=10+8)
Input/split_events_by_not_found-16                                                                        1.592k ± 6%   1.587k ± 5%       ~ (p=0.592 n=10)
…