@efd6 2023-09-04
This change fixes the flakey test in #34929. Applying the same (similar; actually sensibly readable this time) grain of sand as in the [analysis](https://github.com/elastic/beats/issues/34929#issuecomment-1515566040) and unskipping the test now passes, as expected since there can be no race now.

```diff
diff --git a/x-pack/filebeat/input/httpjson/input_test.go b/x-pack/filebeat/input/httpjson/input_test.go
index e88a0a28d3..91ba8a1125 100644
--- a/x-pack/filebeat/input/httpjson/input_test.go
+++ b/x-pack/filebeat/input/httpjson/input_test.go
@@ -374,7 +374,7 @@ var testCases = []struct {
                },
        },
        {
-               skipReason: "flakey test - see https://github.com/elastic/beats/issues/34929",
+               // skipReason: "flakey test - see https://github.com/elastic/beats/issues/34929",

                name: "first_event",
                setupServer: func(t testing.TB, h http.HandlerFunc, config map[string]interface{}) {
diff --git a/x-pack/filebeat/input/httpjson/request.go b/x-pack/filebeat/input/httpjson/request.go
index 6a1d926ab4..3ed904f16d 100644
--- a/x-pack/filebeat/input/httpjson/request.go
+++ b/x-pack/filebeat/input/httpjson/request.go
@@ -16,6 +16,7 @@ import (
        "reflect"
        "strconv"
        "strings"
+       "time"

        "github.com/PaesslerAG/jsonpath"

@@ -741,6 +742,10 @@ func newPublisher(trCtx *transformContext, pub inputcursor.Publisher, publish bo
 }

 func (p *publisher) event(_ context.Context, msg mapstr.M) {
+       if p.n == 0 {
+               time.Sleep(10 * time.Millisecond)
+       }
+
        if p.pub != nil {
                event, err := makeEvent(msg)
                if err != nil {
```

```

$ go test -v -run TestInput/first_event
=== RUN   TestInput
=== RUN   TestInput/first_event
--- PASS: TestInput (2.03s)
    --- PASS: TestInput/first_event (2.03s)
PASS
ok  	github.com/elastic/beats/v7/x-pack/filebeat/input/httpjson	2.841s
```

For completeness, the same change as here applied to main.
```diff
diff --git a/x-pack/filebeat/input/httpjson/input_test.go b/x-pack/filebeat/input/httpjson/input_test.go
index e88a0a28d3..91ba8a1125 100644
--- a/x-pack/filebeat/input/httpjson/input_test.go
+++ b/x-pack/filebeat/input/httpjson/input_test.go
@@ -374,7 +374,7 @@ var testCases = []struct {
                },
        },
        {
-               skipReason: "flakey test - see https://github.com/elastic/beats/issues/34929",
+               // skipReason: "flakey test - see https://github.com/elastic/beats/issues/34929",

                name: "first_event",
                setupServer: func(t testing.TB, h http.HandlerFunc, config map[string]interface{}) {
diff --git a/x-pack/filebeat/input/httpjson/request.go b/x-pack/filebeat/input/httpjson/request.go
index f92d2944c7..2f93d5d391 100644
--- a/x-pack/filebeat/input/httpjson/request.go
+++ b/x-pack/filebeat/input/httpjson/request.go
@@ -16,6 +16,7 @@ import (
        "reflect"
        "strconv"
        "strings"
+       "time"

        "github.com/PaesslerAG/jsonpath"

@@ -706,6 +707,10 @@ func processAndPublishEvents(trCtx *transformContext, events <-chan maybeMsg, pu
                        continue
                }

+               if n == 0 {
+                       time.Sleep(10 * time.Millisecond)
+               }
+
                if publish {
                        event, err := makeEvent(maybeMsg.msg)
                        if err != nil {
```
```
 go test -v -run TestInput/first_event
=== RUN   TestInput
=== RUN   TestInput/first_event
    input_test.go:1313:
        	Error Trace:	.../github.com/elastic/beats/x-pack/filebeat/input/httpjson/input_test.go:1313
        	Error:      	Not equal:
        	            	expected: map[string]interface {}{"first":"a", "foo":"b"}
        	            	actual  : map[string]interface {}{"first":"none", "foo":"b"}

        	            	Diff:
        	            	--- Expected
        	            	+++ Actual
        	            	@@ -1,3 +1,3 @@
        	            	 (map[string]interface {}) (len=2) {
        	            	- (string) (len=5) "first": (string) (len=1) "a",
        	            	+ (string) (len=5) "first": (string) (len=4) "none",
        	            	  (string) (len=3) "foo": (string) (len=1) "b"
        	Test:       	TestInput/first_event
--- FAIL: TestInput (2.03s)
    --- FAIL: TestInput/first_event (2.03s)
FAIL
exit status 1
FAIL	github.com/elastic/beats/v7/x-pack/filebeat/input/httpjson	2.694s```